\name{cgauss}
\alias{cgauss}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Contrast Gaussian expansion algorithm for AVIRIS images}
\description{
  Applies the contrast Gaussian expantion algorithm to de input data.
}
\usage{
cgauss(Z)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Z}{Input data as a matrix or vector}
}

\value{
  Returns the expansion result
}

\author{Marcelo Almiron and Adrian Muract}

\seealso{See Also \code{\link{clineal}}}

\keyword{methods}% at least one, from doc/KEYWORDS

