\name{checkTab}
\alias{checkTab}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to check active tab}
\description{
  Checks the active tab of the GUI.
}
\usage{
checkTab()
}
%- maybe also 'usage' for other objects documented here.

\author{Talita Perciano}

\keyword{methods}% at least one, from doc/KEYWORDS

