\name{contBriImg}
\alias{contBriImg}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Contrast and brightness of an image}
\description{
  Applies values of contrast and brightness on an image.
}
\usage{
contBriImg(img, cont, bri)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{img}{Input image}
  \item{cont}{New value for contrast (0.0 to 1.0)}
  \item{bri}{New value for brightness (-1.0 to 1.0)}
}

\value{
  Return the result image after applying the values of contrast and brightness
}

\author{Talita Perciano}


\keyword{methods}% at least one, from doc/KEYWORDS

