\name{contrast}
\alias{contrast}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Interface to choose the contrast expansion type}
\description{
  Chooses the type of contrast expansion to apply on \code{band}.
}
\usage{
contrast(band, type = c("gauss", "lineal"), ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{band}{One band of the image (matrix)}
  \item{type}{Type of the contrast expansion}
  \item{\dots}{Other possible arguments}
}


\author{Marcelo Almiron and Adrian Muract}

\seealso{See Also \code{\link{cgauss}}, \code{\link{clineal}}}

\keyword{methods}% at least one, from doc/KEYWORDS

