\name{Grey}
\alias{Grey}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Interface for gray scale view for AVIRIS images}
\description{
  Makes an interface for gray scale view of an image band.
}
\usage{
Grey(band, x0, y0, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{band}{One band of the image (matrix)}
  \item{x0}{1}
  \item{y0}{1}
  \item{\dots}{Other possible arguments}
}

\author{Marcelo Almiron and Adrian Muract}

\seealso{See Also \code{\link{plot_band.aviris_band}}, \code{\link{RGB}}}

\keyword{methods}% at least one, from doc/KEYWORDS

