\name{initialize-methods}
\docType{methods}
\alias{initialize-methods}
\alias{initialize,aviris_band-method}
\alias{initialize,aviris_image-method}
\alias{initialize,aviris_scene-method}
\alias{initialize,aviris_training-method}
\title{Methods for Function initialize}
\description{
Methods for function \code{initialize}
}
\section{Methods}{
\describe{

\item{.Object = "aviris_band"}{Create an AVIRIS band.}

\item{.Object = "aviris_image"}{Create an AVIRIS image.}

\item{.Object = "aviris_scene"}{Create an AVIRIS scene.}

\item{.Object = "aviris_training"}{Create a new category for training.}
}}

\keyword{methods}

