\name{lband}
\alias{lband}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Interface to load an AVIRIS image band}
\description{
  Loads an image band of a scene.
}
\usage{
lband(scene, b)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{scene}{An object of type scene}
  \item{b}{Number of the band}
}

\author{Marcelo Almiron and Adrian Muract}

\seealso{See Also \code{\link{lbandsample}}}

\keyword{methods}% at least one, from doc/KEYWORDS

