\name{lbandsample}
\alias{lbandsample}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Interface to load an AVIRIS image band sample}
\description{
  Loads an image band sample of a scene.
}
\usage{
lbandsample(scene, b)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{scene}{An object of type scene}
  \item{b}{Number of the band}
}

\author{Marcelo Almiron and Adrian Muract}

\seealso{See Also as \code{\link{lband}}.}

\keyword{methods}% at least one, from doc/KEYWORDS

