\name{limage}
\alias{limage}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Interface to load the header of an AVIRIS image}
\description{
  Loads the header of an image.
}
\usage{
limage(H, type)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{H}{Name of the header}
  \item{type}{Type of the image ("reflectance" or "radiance")}
}

\author{Marcelo Almiron and Adrian Muract}


\keyword{methods}% at least one, from doc/KEYWORDS

