\name{loadBand}
\alias{loadBand}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to load an AVIRIS image band}
\description{
  Loads an image band using \code{lband} interface.
}
\usage{
loadBand(I, X = 5, C = 614, F = 512, B = 224)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{I}{Name of the image file}
  \item{X}{Number of the band}
  \item{C}{Number of columns of each band}
  \item{F}{Number of lines of each band}
  \item{B}{Total number of bands}
}

\author{Marcelo Almiron and Adrian Muract}

\seealso{See Also as \code{\link{loadBandSample}}}

\keyword{methods}% at least one, from doc/KEYWORDS

