\name{lscene}
\alias{lscene}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Interfece to load an AVIRIS image scene}
\description{
  Loads an AVIRIS image scene.
}
\usage{
lscene(image, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{image}{Name of the file}
  \item{n}{Number of the scene}
}

\author{Marcelo Almiron and Adrian Muract}


\keyword{methods}% at least one, from doc/KEYWORDS

