\name{medianImg}
\alias{medianImg}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Median filter}
\description{
  Applies the median filter on an image
}
\usage{
medianImg(img, mask)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{img}{The input image as a matrix}
  \item{mask}{The mask length}
}

\author{Talita Perciano}


\keyword{methods}% at least one, from doc/KEYWORDS

