\name{modalDialog}
\alias{modalDialog}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Modal dialog}
\description{
  Builds a tcltk modal dialog.
}
\usage{
modalDialog(title, question, entryInit, entryWidth = 20, returnValOnCancel = "ID_CANCEL")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{title}{Title of the dialog}
  \item{question}{Question of the dialog}
  \item{entryInit}{The initial entry value}
  \item{entryWidth}{The entry width}
  \item{returnValOnCancel}{Value to be returned on cancel}
}

\author{Talita Perciano}


\keyword{methods}% at least one, from doc/KEYWORDS

