\name{print.imagematrix}
\alias{print.imagematrix}
\title{Print information on a given imagematrix object}
\description{
  This function outputs information on a given imagematrix object.
}
\usage{
  \method{print}{imagematrix}(x, ...)}
\arguments{
  \item{x}{target image}
  \item{...}{ignored (dummy)}
}
\examples{
  data(logo)
  print(logo)
}
\seealso{\code{\link{imagematrix}}}
\keyword{misc}


