\name{print_information.aviris_band}
\alias{print_information.aviris_band}

\title{Function to print an AVIRIS band}
\description{
  Prints information about an AVIRIS band.
}

\usage{
	\method{print_information}{aviris_band}(Object)
}

\arguments{
  \item{Object}{
	An object of class aviris_band.
  }
}

\author{Marcelo Almiron and Adrian Muract}


\keyword{methods}% at least one, from doc/KEYWORDS

