\name{print_information.aviris_image}
\alias{print_information.aviris_image}

\title{Function to print an AVIRIS image}
\description{
  Prints information about an AVIRIS image.
}

\usage{
\method{print_information}{aviris_image}(Object)
}

\arguments{
  \item{Object}{
	An object of class aviris_image.
  }
}

\author{Marcelo Almiron and Adrian Muract}


\keyword{methods}% at least one, from doc/KEYWORDS



