\name{print_information.aviris_scene}
\alias{print_information.aviris_scene}

\title{Function to print an AVIRIS scene}

\description{
  Prints information about an AVIRIS scene.
}

\usage{
	\method{print_information}{aviris_scene}(Object)
}

\arguments{
  \item{Object}{An object of class aviris_scene.}
}

\author{Marcelo Almiron and Adrian Muract}

\keyword{methods}% at least one, from doc/KEYWORDS

