\name{print_information.aviris_training}
\alias{print_information.aviris_training}

\title{Function to print an AVIRIS training category}

\description{
  Prints information about an AVIRIS training category.
}

\usage{
	\method{print_information}{aviris_training}(Object)
}

\arguments{
  \item{Object}{An object of class aviris_training.}
}

\author{Marcelo Almiron and Adrian Muract}

\keyword{methods}% at least one, from doc/KEYWORDS

