\name{read.aviris}
\alias{read.aviris}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to read an AVIRIS image}
\description{
  Reads an AVIRIS image.
}
\usage{
read.aviris(fileName, bandsIndexes, bands_local, use_parallel)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fileName}{The name of the file}
  \item{bandsIndexes}{Indexes of the bands to be read}
  \item{bands_local}{Local variable to store the bands}
  \item{use_parallel}{Indicates if the function should be run in parallel or not (1=yes, 0=no)}
}

\author{Talita Perciano}

\seealso{See Also as \code{\link{read.lan}}, ~~~ }

\keyword{methods}% at least one, from doc/KEYWORDS

