\name{read.lan}
\alias{read.lan}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to read a LAN image}
\description{
  Reads a LAN image.
}
\usage{
read.lan(arquivo)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{arquivo}{The name of the file}
}

\author{Talita Perciano}


\keyword{methods}% at least one, from doc/KEYWORDS

