\name{RGB}
\alias{RGB}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Interface for RGB view for AVIRIS images}
\description{
  Makes an interface for RGB view of an image band.
}
\usage{
RGB(red, green, blue, x0, y0, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{red}{The R band}
  \item{green}{The G band}
  \item{blue}{The B band}
  \item{x0}{1}
  \item{y0}{1}
  \item{\dots}{Other possible arguments}

}

\author{Marcelo Almiron and Adrian Muract}

\seealso{See Also as \code{\link{Grey}}.}

\keyword{methods}% at least one, from doc/KEYWORDS

