\name{ripa-package}
\alias{ripa-package}
\alias{ripa}
\docType{package}
\title{
R Image Processing and Analysis
}
\description{
A package including various functions for image processing and analysis. With this package is possible to process and analyse RGB, LAN (multispectral) and AVIRIS (hiperspectral) images. This packages also provides functions for reading JPEG files, extracted from the archived 'rimage' package.
}
\details{
\tabular{ll}{
Package: \tab ripa\cr
Type: \tab Package\cr
Version: \tab 2.0-1\cr
Date: \tab 2014-05-12\cr
License: \tab GPL version 2 or newer\cr
}
}
\author{
Talita Perciano and Alejandro C Frery

Maintainer: Talita Perciano <talitaperciano@gmail.com>
}

\keyword{ package }


