\name{RIPAgui}
\alias{RIPAgui}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to build the ripa package GUI}
\description{
  Builds the GUI for ripa package. After loading the package, users should run this function to use all the developed tools.
}
\usage{
RIPAgui()
}
%- maybe also 'usage' for other objects documented here.

\author{Talita Perciano}


\keyword{methods}% at least one, from doc/KEYWORDS

