\name{stretchImg}
\alias{stretchImg}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to apply contrast linear stretch}
\description{
  Applies contrast linear stretch to an image.
}
\usage{
stretchImg(img)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{img}{Input image as a matrix}
}

\author{Talita Perciano}


\keyword{methods}% at least one, from doc/KEYWORDS

