\name{takeSamples}
\alias{takeSamples}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to take training samples for AVIRIS images}
\description{
  Take samples for future training.
}
\usage{
takeSamples(t, n = NULL, Sample = NULL, Line = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{t}{Training class}
  \item{n}{Number of samples}
  \item{Sample}{Column of the image}
  \item{Line}{Line of the image}
}

\author{Marcelo Almiron and Adrian Muract}


\keyword{methods}% at least one, from doc/KEYWORDS

