\name{wband}
\alias{wband}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Interface to save an AVIRIS image band}
\description{
  Saves an image band of a scene.
}
\usage{
wband(scene, band)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{scene}{An object of type scene}
  \item{band}{Number of the band}
}

\author{Marcelo Almiron and Adrian Muract}

\seealso{See Also as \code{\link{lband}}.}

\keyword{methods}% at least one, from doc/KEYWORDS

