\name{write.lan}
\alias{write.lan}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to save LAN images}
\description{
  Saves LAN images.
}
\usage{
write.lan(arquivo, img)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{arquivo}{Name of the file}
  \item{img}{Data to be saved}
}

\author{Talita Perciano}

\seealso{See Also \code{\link{read.lan}}}

\keyword{methods}% at least one, from doc/KEYWORDS

