\name{zoom}
\alias{zoom}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Interface for zoomGrey and zoomRGB}
\description{
	Interface for zoomGrey and zoomRGB.
}
\usage{
zoom(R = NULL, G = NULL, B = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{R}{The R band}
  \item{G}{The G band}
  \item{B}{The B band}
}

\author{Marcelo Almiron and Adrian Muract}

\seealso{See Also \code{\link{zoomRGB}}, \code{\link{zoomGrey}}}

\keyword{methods}% at least one, from doc/KEYWORDS

