\name{zoomGrey}
\alias{zoomGrey}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to apply zoom to grey images}
\description{
  Applies zoom to an grey image.
}
\usage{
zoomGrey(band)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{band}{Data input}
}

\author{Marcelo Almiron and Adrian Muract}

\seealso{See Also \code{\link{zoomRGB}}}

\keyword{methods}% at least one, from doc/KEYWORDS

