\name{zoomRGB}
\alias{zoomRGB}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to apply zoom to RGB images}
\description{
  Applies zoom to a RGB image.
}
\usage{
zoomRGB(Red, Green, Blue)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Red}{The R band}
  \item{Green}{The G band}
  \item{Blue}{The B band}
}

\author{Marcelo Almiron and Adrian Muract}

\seealso{See Also \code{\link{zoomGrey}}}

\keyword{methods}% at least one, from doc/KEYWORDS

