\name{Zprofile}
\alias{Zprofile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function for Z Profile}
\description{
  Show the Z Profile of a pixel.
}
\usage{
Zprofile(scene, X = NULL, Y = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{scene}{Scene}
  \item{X}{X-coordinate}
  \item{Y}{Y-coordinate}
}

\author{Marcelo Almiron and Adrian Muract}


\keyword{methods}% at least one, from doc/KEYWORDS

