\name{read.aviris}
\alias{read.aviris}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to read an AVIRIS image}
\description{
  Reads an AVIRIS image.
}
\usage{
read.aviris(fileName)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fileName}{The name of the file}
}

\author{Talita Perciano}

\seealso{See Also as \code{\link{read.lan}}, ~~~ }

\keyword{methods}% at least one, from doc/KEYWORDS

