\name{ripa-package}
\alias{ripa-package}
\alias{ripa}
\docType{package}
\title{
R Image Processing and Analysis
}
\description{
A package including various functions for image processing and analysis of JPG, LAN and AVIRIS images. The package contains many tools for statistical analysis of multispectral and hyperspectral images.
}
\details{
\tabular{ll}{
Package: \tab ripa\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2009-07-17\cr
License: \tab GPL version 2 or newer\cr
}
}
\author{
Talita Perciano and Alejandro C Frery

Maintainer: Talita Perciano <talitaperciano@gmail.com>
}

\keyword{ package }


